fix: check for presigned url
In my previous commit, I fixed the regex that did not capture s3 bucket url structure and failed to distinguish presigned urls. However, I realized that just fixing the regex is not enough as now it does not distinguish between public and presigned s3 buckets. so I introduced an improved check that only filters presigned buckets.